Fix bug in computing cell type for array classes
The bug:  Inner array classes didn't work properly because the
'computeCellType' method assumed that the members of the array class
were already entered.  When this method is called during member entry,
that may not be true. I put in a test to disable the method when the
members aren't there. This seems to work!  Apparently the method is
being called later too. More investigation is needed.